+++ backend/src/main/java/com/beyond/specguard/auth/exception/errorcode/AuthErrorCode.java
@@ -35,7 +35,9 @@ public enum AuthErrorCode implements ErrorCode {
     DUPLICATE_SLUG(HttpStatus.CONFLICT, "DUPLICATE_SLUG", "이미 사용 중인 슬러그입니다."),
     USER_NOT_FOUND(HttpStatus.NOT_FOUND, "USER_NOT_FOUND", "사용자를 찾을 수 없습니다."),
     DUPLICATE_COMPANY_NAME(HttpStatus.CONFLICT,"DUPLICATE_COMPANY_NAME" ,"이미 사용중인 회사명입니다" ), 
-    DUPLICATE_EMAIL_IN_COMPANY(HttpStatus.CONFLICT,"DUPLICATE_EMAIL_IN_COMPANY","이미 내부에서 사용중인 이메일 입니다" );
+    DUPLICATE_EMAIL_IN_COMPANY(HttpStatus.CONFLICT,"DUPLICATE_EMAIL_IN_COMPANY","이미 내부에서 사용중인 이메일 입니다" ),
+    SESSION_CONFLICT(HttpStatus.CONFLICT, "DUPLICATE_SESSION", "사용중인 세션입니다."),
+    INTERNAL_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "INTERNAL_ERROR", "내부 에러가 발생했습니다.");
 
     private final HttpStatus status;
     private final String code;
