+++ backend/src/main/java/com/beyond/specguard/auth/model/filter/JwtFilter.java
@@ -7,11 +7,11 @@ import com.beyond.specguard.auth.exception.AuthException;
 import com.beyond.specguard.auth.exception.errorcode.AuthErrorCode;
 import com.beyond.specguard.auth.model.entity.ClientUser;
 import com.beyond.specguard.auth.model.repository.ClientUserRepository;
-import com.beyond.specguard.auth.model.service.common.RedisTokenService;
-import com.beyond.specguard.auth.model.service.local.CustomUserDetails;
+import com.beyond.specguard.auth.model.service.CustomUserDetails;
+import com.beyond.specguard.auth.model.service.RedisTokenService;
 import com.beyond.specguard.auth.model.token.AdminAuthenticationToken;
 import com.beyond.specguard.auth.model.token.ClientAuthenticationToken;
-import com.beyond.specguard.common.util.JwtUtil;
+import com.beyond.specguard.common.jwt.JwtUtil;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
