+++ backend/src/main/java/com/beyond/specguard/auth/model/handler/CustomSuccessHandler.java
@@ -1,10 +1,10 @@
-package com.beyond.specguard.auth.model.handler.local;
+package com.beyond.specguard.auth.model.handler;
 
 import com.beyond.specguard.admin.model.service.InternalAdminDetails;
-import com.beyond.specguard.auth.model.service.common.RedisTokenService;
-import com.beyond.specguard.auth.model.service.local.CustomUserDetails;
+import com.beyond.specguard.auth.model.service.CustomUserDetails;
+import com.beyond.specguard.auth.model.service.RedisTokenService;
+import com.beyond.specguard.common.jwt.JwtUtil;
 import com.beyond.specguard.common.util.CookieUtil;
-import com.beyond.specguard.common.util.JwtUtil;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
