+++ backend/src/main/java/com/beyond/specguard/auth/model/handler/oauth2/OAuth2SuccessHandler.java
@@ -1,8 +1,8 @@
 package com.beyond.specguard.auth.model.handler.oauth2;
 
+import com.beyond.specguard.auth.model.service.RedisTokenService;
 import com.beyond.specguard.auth.model.service.oauth2.CustomOAuth2UserDetails;
-import com.beyond.specguard.auth.model.service.common.RedisTokenService;
-import com.beyond.specguard.common.util.JwtUtil;
+import com.beyond.specguard.common.jwt.JwtUtil;
 import com.beyond.specguard.common.properties.AppProperties;
 import com.beyond.specguard.common.util.CookieUtil;
 import jakarta.servlet.http.Cookie;
